+++ assets/javascripts/kitten/components/molecules/cards/project-card/stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ProjectCard } from './index'
 import { DocsPage } from 'storybook/docs-page'
-import { Text } from 'kitten'
+import { Text, Button } from 'kitten'
 
 export default {
   title: 'Molecules/Cards/ProjectCard',
@@ -20,12 +20,10 @@ export default {
 
 const argTypes = {
   status: { control: 'select' },
-  sticker: { control: 'text' },
   progressProps: { control: 'object' },
 }
 
 const args = {
-  sticker: 'Prolongation',
   status: 'normal',
   imageProps: {
     src: `/kitten-${Math.floor(Math.random() * 10)}.jpg`,
@@ -41,11 +39,16 @@ const args = {
 
 export const Default = ({ progressProps, ...args }) => (
   <ProjectCard {...args}>
+    <ProjectCard.Sticker
+      className="k-u-transform-capitalize k-u-hidden@xs-down"
+    >
+      Prolongation
+    </ProjectCard.Sticker>
     <ProjectCard.Title>
       The Office, la série culte décortiquée par S!CK
     </ProjectCard.Title>
     <ProjectCard.Line>
-      <Text size="micro" fontStyle="italic">
+      <Text size="micro">
         par{' '}
         <a href="#" className="k-u-link k-u-link-font1 k-u-weight-bold">
           Bidules
@@ -85,6 +88,62 @@ export const Default = ({ progressProps, ...args }) => (
 Default.args = args
 Default.argTypes = argTypes
 
+export const WithButton = ({ progressProps, ...args }) => (
+  <ProjectCard {...args}>
+    <ProjectCard.Sticker
+      className="k-u-transform-capitalize k-u-hidden@xs-down"
+    >
+      Prolongation
+    </ProjectCard.Sticker>
+    <ProjectCard.Title>
+      The Office, la série culte décortiquée par S!CK
+    </ProjectCard.Title>
+    <ProjectCard.Line>
+      <Text size="micro">
+        par{' '}
+        <a href="#" className="k-u-link k-u-link-font1 k-u-weight-bold">
+          Bidules
+        </a>
+      </Text>
+    </ProjectCard.Line>
+    <ProjectCard.ItemsLine>
+      <ProjectCard.Item>
+        <Text className="k-u-ellipsis" weight="bold" size="tiny">
+          134
+        </Text>
+        <Text className="k-u-ellipsis" size="micro">
+          contributeurs
+        </Text>
+      </ProjectCard.Item>
+      <ProjectCard.Item>
+        <Text className="k-u-ellipsis" weight="bold" size="tiny">
+          7 jours
+        </Text>
+        <Text className="k-u-ellipsis" size="micro">
+          restants
+        </Text>
+      </ProjectCard.Item>
+      <ProjectCard.Item>
+        <Text className="k-u-ellipsis" weight="bold" size="tiny">
+          9 930 €
+        </Text>
+        <Text className="k-u-ellipsis" size="micro">
+          sur 12 000 €
+        </Text>
+      </ProjectCard.Item>
+    </ProjectCard.ItemsLine>
+    <ProjectCard.Progress {...progressProps} />
+    <ProjectCard.Line>
+      <Button fit="fluid" modifier="helium" size="tiny">
+        Participer
+      </Button>
+    </ProjectCard.Line>
+  </ProjectCard>
+)
+
+WithButton.args = args
+WithButton.argTypes = argTypes
+
 export const AsLink = args => <Default {...args} />
 
 AsLink.args = {
@@ -114,3 +173,23 @@ WithVideo.args = {
   ],
 }
 WithVideo.argTypes = argTypes
+
+export const Loading = args => (
+  <ProjectCard {...args}>
+    <ProjectCard.Title />
+    <ProjectCard.Line />
+    <ProjectCard.ItemsLine>
+      <ProjectCard.Item />
+      <ProjectCard.Item />
+      <ProjectCard.Item />
+    </ProjectCard.ItemsLine>
+    <ProjectCard.Progress />
+  </ProjectCard>
+)
+
+Loading.args = {
+  ...args,
+  imageProps: {},
+  loading: true,
+}
+Loading.argTypes = argTypes